pyproject.toml
[project]
name = "wccomps-bot"
version = "1.0.0"
description = "Discord bot and web interface for WCComps team management"
requires-python = ">=3.11"
dependencies = [
    # Discord bot
    "discord.py>=2.4.0",
    "python-dotenv>=1.0.1",

    # Django web app
    "django>=5.0,<5.1",
    "django-allauth>=0.57",
    "pyjwt>=2.8.0",
    "gunicorn>=21.2",
    "whitenoise>=6.6",

    # Authentik integration
    "authlib>=1.3.1",
    "authentik-client>=2024.12.1",

    # Shared dependencies
    "psycopg2-binary>=2.9.9",
    "requests>=2.32.3",
    "aiohttp>=3.9",
    "xkcdpass>=1.19.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-django>=4.7",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot", "web"]

[dependency-groups]
dev = [
    "ruff>=0.14.0",
    "pytest>=7.4",
    "pytest-django>=4.7",
    "pytest-asyncio>=0.21",
    "pylint>=3.0",
    "pylint-django>=2.6",
    "mypy>=1.13",
    "django-stubs>=5.1.1",
    "types-requests>=2.32",
]

[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = "bot.tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "web.tests.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "wccomps.settings"

[tool.mutmut]
paths_to_mutate = ["bot/discord_manager.py"]
tests_dir = ["../bot/tests"]
runner = "export DJANGO_SETTINGS_MODULE=wccomps.settings PYTHONPATH=../web:.. && pytest ../bot/tests -x -q"
debug = true
